# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gestão de EPIs.",
        "title": "SaaS EPI API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Suporte API",
            "url": "http://www.radaptech.com.br",
            "email": "suporte@seusaas.com.br"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/epis": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna uma lista paginada de EPIs",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EPIs"
                ],
                "summary": "Listar EPIs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Página",
                        "name": "pagina",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Quantidade por página",
                        "name": "quantidade",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.EpiPaginado"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Adiciona um novo EPI ao estoque da empresa",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EPIs"
                ],
                "summary": "Cadastrar novo EPI",
                "parameters": [
                    {
                        "description": "Dados do EPI",
                        "name": "epi",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.EpiInserir"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "EPI cadastrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "CA já registrado",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    },
                    "422": {
                        "description": "Erro de validação (Data, Integridade)",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    }
                }
            }
        },
        "/epis/dashboard": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna dados resumidos para o dashboard de EPIs",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EPIs"
                ],
                "summary": "Dashboard de EPIs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.EpiDashBord"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    }
                }
            }
        },
        "/epis/funcionario/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna a lista de EPIs vinculados a um funcionário específico",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EPIs"
                ],
                "summary": "EPIs por funcionário",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do Funcionário",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.EpiDtoDevolucao"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    }
                }
            }
        },
        "/epis/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAut": []
                    }
                ],
                "description": "Retorna os detalhes de um EPI específico",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EPIs"
                ],
                "summary": "Buscar EPI por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do EPI",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.EpiDto"
                        }
                    },
                    "422": {
                        "description": "EPI não encontrado",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Atualiza os dados de um EPI existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EPIs"
                ],
                "summary": "Atualizar EPI",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do EPI",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados para atualização",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateEpiInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "CA já cadastrado",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove um EPI do cadastro",
                "tags": [
                    "EPIs"
                ],
                "summary": "Deletar EPI",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do EPI",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sem conteúdo"
                    },
                    "404": {
                        "description": "EPI não encontrado",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "helper.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.EpiDashBord": {
            "type": "object",
            "properties": {
                "alerta_minimo": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                }
            }
        },
        "model.EpiDto": {
            "type": "object",
            "properties": {
                "alerta_minimo": {
                    "type": "integer"
                },
                "ca": {
                    "type": "string"
                },
                "descricao": {
                    "type": "string"
                },
                "fabricante": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                },
                "protecao": {
                    "$ref": "#/definitions/model.TipoProtecaoDto"
                },
                "tamanhos": {
                    "description": "Slive de objetos Tamanho (id e nome)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.TamanhoDto"
                    }
                },
                "validade_ca": {
                    "type": "string"
                }
            }
        },
        "model.EpiDtoDevolucao": {
            "type": "object",
            "properties": {
                "alerta_minimo": {
                    "type": "integer"
                },
                "ca": {
                    "type": "string"
                },
                "descricao": {
                    "type": "string"
                },
                "fabricante": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                },
                "protecao": {
                    "$ref": "#/definitions/model.TipoProtecaoDto"
                },
                "saldoAtual": {
                    "type": "integer"
                },
                "tamanhos": {
                    "description": "Slive de objetos Tamanho (id e nome)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.TamanhoDto"
                    }
                },
                "validade_ca": {
                    "type": "string"
                }
            }
        },
...
        "model.EpiInserir": {
            "type": "object",
            "required": [
                "alerta_minimo",
                "ca",
                "data_validade_ca",
                "fabricante",
                "id_protecao",
                "id_tamanho",
                "nome"
            ],
            "properties": {
                "alerta_minimo": {
                    "type": "integer",
                    "minimum": 0
                },
                "ca": {
                    "type": "string",
                    "maxLength": 20
                },
                "data_validade_ca": {
                    "type": "string"
                },
                "descricao": {
                    "type": "string",
                    "maxLength": 250
                },
                "fabricante": {
                    "type": "string",
                    "maxLength": 100
                },
                "id_protecao": {
                    "type": "integer"
                },
                "id_tamanho": {
                    "description": "Lista de tamanhos permitidos para este EPI",
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "integer"
                    }
                },
                "nome": {
                    "type": "string"
                }
            }
        },
...
        "model.TamanhoDto": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "tamanho": {
                    "type": "string"
                }
            }
        },
...
        "model.TipoProtecaoDto": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                }
            }
        },
        "model.UpdateEpiInput": {
            "type": "object",
            "properties": {
                "alerta_minimo": {
                    "type": "integer"
                },
                "ca": {
                    "type": "string"
                },
                "descricao": {
                    "type": "string"
                },
                "fabricante": {
                    "type": "string"
                },
                "id_protecao": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                },
                "tamanhos": {
                    "description": "Se enviado, você deve resetar os tamanhos no banco",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "validade_ca": {
                    "type": "string"
                }
            }
        },
...
        "service.EpiPaginado": {
            "type": "object",
            "properties": {
                "epis": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.EpiDto"
                    }
                },
                "pagina": {
                    "type": "integer",
                    "format": "int32"
                },
                "paginaFinal": {
                    "type": "integer",
                    "format": "int32"
                },
                "total": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}